platform/x86: dell-smo8800: Allow i2c_client instantiation without IRQ
On the Dell XPS 15 9550 the SMO8810 ACPI device may be disabled,
while the accelerometer is always there. I suspect this happens when
the XPS 15 9550 has the big (90Wh) battery which covers the room
for a 2.5" sata disk, so the freefall function is disabled in the BIOS.

The ACPI device having its status set to "not present" is worked
around by a quirk in drivers/acpi/x86/utils.c marking it as always
present. But when the function is disabled in the BIOS the IRQ is
not setup so the platform_get_irq() call fails.

This means no /dev/freefall functionality, but the accelerometer
can still be used as an accelerometer.

Allow i2c_client instantiation to continue without an IRQ and
only fail probe() if there is both no i2c_client instantiated
and no IRQ.

Note that IRQ 0 is not a valid IRQ number for platform IRQs
and this patch relies on that.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>