media: atomisp: Add t4ka3 driver
Add the t4ka3 driver from:
https://github.com/kitakar5525/surface3-atomisp-cameras.git

With the following bits removed:
1. The VCM and VCM-OTP support removed since the mainline kernel models
   VCMs as separate v4l2-subdev-s.
2. The integration-factor (t4ka3_get_intg_factor()) this provided info
   to userspace for an atomisp private IOCTL which has been removed.

This is added as atomisp specific sensor driver for now. Since the atomisp
code can use generic v4l2 sensor drivers now this is an intermediate step
towards transforming this into a generic sensor driver and adding it as
such.

Note this does not work yet on devices with a Whisky Cove PMIC like
the Xiaomi Mipad2 since it uses atomisp_gmin_platform.c functions
for power-management and those do not support Whisky Cove PMICs.

To fix this this needs to be converted to use normal ACPI pm in
combination with runtime-pm support.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>