media: atomisp: t4ka3: Add _REG to all register address defines
Add _REG to all register address defines and sort them all by register
address.

This is a preparation patch for replacing the pixels_per_line and
lines_per_frame variables with defines (since these are the same for all
modes). Adding _REG to the register defines makes it clear that the to
be added T4K3A_PIXELS_PER_LINE and T4K3A_LINES_PER_FRAME defines are
not register addresses.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>