media: atomisp: t4ka3: Move writing of registers from set_fmt() to s_stream(sd, 1)
With the switch to runtime-pm, the sensor is only turned on at stream-on
time, so trying to write the registers from set_fmt() will fail.

Move the writing of the mode registers to s_stream() after
the pm_runtime_get_sync() call.

Since the sensor is not on before s_stream(sd, 1) is called, the recovery
code which power-cycles the sensor is no longer necessary as the sensor
already gets power-cycled on stream on now. Remove the recovery code.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>