media: atomisp: t4ka3: Do not write ctrl values to sensor while not streaming
With the switch to runtime-pm, the sensor is only turned on at stream-on
time, so trying to write the registers when ctrls are set while not
streaming will fail.

Change t4ka3_s_ctrl() to not write the registers when not streaming and
make s_stream(sd, 1) restore all cached control values after
the pm_runtime_get_sync() call has turned the sensor on.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>