media: ipu-bridge: Add Toshiba T4KA3 HID to list of supported sensor
The Xiaomi Mi Pad 2 tablet uses a Toshiba T4KA3 sensor, using
a Xiaomi specific Hardware-ID of "XMMC0003" add this to the sensor-list
of the ipu-bridge code.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Cc: Kate Hsuan <hpa@redhat.com>
Kate Hsuan and I have been working on a driver for this sensor and we
have things working now. The t4ka3 sensor driver will be posted upsteam
soon (once we have done a bunch of cleanup work and added hwcfg validation)