platform/x86: int3472: Map GPIO type 0x12 to reset
GPIO type 0x12 is a new type encountered on some Meteor Lake laptops
it is supposed to be the handshake signal for the not yet supported
CVS / Lattice MIPI Aggregator chip used on some devices.

But it seemed that at least on some HP Elite models GPIO type 0x12
is used without a CVS chip.

Simply map it to reset to get things working for now. This may needs
to be revisited when CVS support lands.

Note most sensor drivers already support the presence of a reset GPIO.

Cc: Hao Yao <hao.yao@intel.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>